init commit with addingspmd kernel support in partial compute
